+++ src/main/java/sws/songpin/domain/follow/dto/response/FollowingListResponseDto.java
@@ -1,5 +1,7 @@
 package sws.songpin.domain.follow.dto.response;
 
+import sws.songpin.domain.follow.dto.FollowDto;
+
 import java.util.List;
 
 public record FollowingListResponseDto(
